Fix terminal display issue
